some refactoring changes to satisfy mypy.
